django/requirements.txt
Django>=3.0
mysqlclient
djangorestframework
django-cors-headers